+++ src/main/java/org/tricol/supplierchain/entity/Produit.java
@@ -4,6 +4,7 @@ package org.tricol.supplierchain.entity;
 import jakarta.persistence.*;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -27,10 +28,10 @@ public class Produit {
     private Double prixUnitaire; 
 
     @Column(name = "stock_actuel", nullable = false)
-    private int stockActuel; 
+    private BigDecimal stockActuel;
 
     @Column(name = "point_commande", nullable = false)
-    private int pointCommande;
+    private BigDecimal pointCommande;
 
     @Column(name = "unite_mesure", nullable = false)
     private String uniteMesure;
